requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
praw>=7.7.0
textblob>=0.17.0
pyyaml>=6.0
